% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_list_time_periods.R
\name{sc_list_time_periods}
\alias{sc_list_time_periods}
\title{Liste der verfuegbaren Zeitperioden fuer ein STATcube time_valueset}
\usage{
sc_list_time_periods(time_valueset_id)
}
\arguments{
\item{time_valueset_id}{ID des time_valueset}
}
\value{
data.frame mit Spalten: id, label, position, start, end (falls vorhanden)
}
\description{
Liste der verfuegbaren Zeitperioden fuer ein STATcube time_valueset
}
