% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_update_json.R
\name{sc_update_json_abfrage}
\alias{sc_update_json_abfrage}
\title{Update der JSON-Datei für eine STATcube-Abfrage mit aktuellsten Daten}
\usage{
sc_update_json_abfrage(
  json_path,
  time_valueset_id,
  time_field_id,
  new_file = TRUE
)
}
\arguments{
\item{json_path}{Pfad der originalen JSON-Datei}

\item{time_valueset_id}{ID des time_valueset}

\item{time_field_id}{ID des time_field}

\item{new_file}{Neue JSON-Datei erstellen oder alte überschreiben}
}
\value{
Pfad der neuen JSON-Datei falls new_file = TRUE
}
\description{
Update der JSON-Datei für eine STATcube-Abfrage mit aktuellsten Daten
}
